requirements.txt
PyMuPDF==1.24.5
google-genai>=0.8.0
numpy==1.26.4
python-dotenv==1.0.1
flask>=3.0.0
